220716 Moon
- myPage.jsp > 주소, 비밀번호 업데이트 기능 추가
- member-mapper.xml > updateInfo 수정 / originPwdCheck, infoUpdatePwd 추가
- Member.java > 위도, 경도 int -> double로 수정
- MemberController, MemberService, MemberServiceImpl, MemberDao > 비밀번호 수정 메소드 추가
- MemberController > 주소 업데이트 메소드 수정

+++ Catchmind/src/main/java/com/e1i4/catchmind/member/model/vo/Member.java
@@ -34,6 +34,6 @@ public class Member {
 	private String partner; //PARTNER	NUMBER
 	private int status; //STATUS	NUMBER
 	private String pic; //PIC	VARCHAR2(100 BYTE)
-	private int longitude; //LONGITUDE	NUMBER
-	private int latitude; //LATITUDE	NUMBER
+	private double longitude; //LONGITUDE	NUMBER
+	private double latitude; //LATITUDE	NUMBER
 }
